feat: disable Pagefind on specific pages
Update layout components to include a `disablePagefind` prop, 
allowing for the disabling of Pagefind functionality on certain 
pages. This change is applied to blog category, tag, and search 
pages to improve performance and user experience by preventing 
unnecessary indexing. 

Additionally, implement a script on the 
search page to pre-fill the search input based on query parameters.